src/integrations/jira_integration.py
import aiohttp
import json
from typing import Dict, Any, List, Optional
from datetime import datetime

from .base_integration import BaseIntegration, SyncResult, TestResult

class JiraIntegration(BaseIntegration):
    """Jira integration for issue tracking and project management."""

    def __init__(self, integration):
        super().__init__(integration)
        self.required_config_fields = ["server_url", "projects"]

    @property
    def api_base_url(self) -> str:
        """Get the API base URL from configuration."""
        server_url = self.integration.config.get("server_url", "")
        return f"{server_url.rstrip('/')}/rest/api/2"

    async def test_connection(self) -> TestResult:
        """Test connection to Jira API."""
        try:
            headers = self.get_auth_headers()

            async with aiohttp.ClientSession() as session:
                async with session.get(
                    f"{self.api_base_url}/myself",
                    headers=headers,
                    timeout=aiohttp.ClientTimeout(total=30)
                ) as response:

                    if response.status == 200:
                        data = await response.json()
                        return TestResult(
                            success=True,
                            message="Connection successful",
                            details={"user": data.get("displayName")}
                        )
                    else:
                        return TestResult(
                            success=False,
                            message=f"HTTP {response.status}"
                        )

        except Exception as e:
            return TestResult(
                success=False,
                message=f"Connection failed: {str(e)}"
            )

    async def sync(self) -> SyncResult:
        """Sync data from Jira projects."""
        # Implementation would sync issues, projects, etc.
        # This is a placeholder implementation

        return SyncResult(
            success=True,
            records_processed=0,
            metadata={"implementation": "placeholder"}
        )

    async def process_webhook(self, event_type: str, payload: Dict[str, Any]) -> bool:
        """Process Jira webhook events."""
        self.log_webhook_received(event_type)

        # Implementation would handle jira:issue_created, jira:issue_updated events
        # This is a placeholder implementation

        return True