requirements.txt
# Core dependencies
boto3==1.28.57
botocore==1.31.57

# AWS CDK
aws-cdk-lib==2.100.0
constructs>=10.0.0

# API & Web
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
httpx==0.25.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0

# AI/ML
anthropic==0.7.1
langchain==0.0.335
openai==1.3.0  # For embeddings

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1
pymongo==4.5.0

# Utils
python-dotenv==1.0.0
PyYAML==6.0.1
structlog==23.2.0
tenacity==8.2.3
prometheus-client==0.23.1
python-json-logger==3.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
moto==4.2.7  # AWS mocking
responses==0.24.1
